+++ test/consumer/echoBytes.typed.sql
@@ -1,2 +1,1 @@
--- @name echoBytes
 select :payloads::bytea[] as payloads